Update the signature of init function to include whether the device is am: 3e7b35ee18
Original change: https://googleplex-android-review.googlesource.com/c/platform/hardware/libhardware/+/12308134

+++ modules/camera/3_4/static_properties_test.cpp
@@ -107,7 +107,7 @@ class StaticPropertiesTest : public Test {
       stream_addresses.push_back(&streams[i]);
     }
     camera3_stream_configuration_t config = {
-        static_cast<uint32_t>(stream_addresses.size()),
+        stream_addresses.size(),
         stream_addresses.data(),
         CAMERA3_STREAM_CONFIGURATION_NORMAL_MODE};
     PrepareDefaultDUT();
@@ -443,7 +443,7 @@ TEST_F(StaticPropertiesTest, ConfigureEmptyStreams) {
 TEST_F(StaticPropertiesTest, ConfigureNullStreams) {
   std::vector<camera3_stream_t*> streams(2, nullptr);
   camera3_stream_configuration_t config = {
-      static_cast<uint32_t>(streams.size()), streams.data(), CAMERA3_STREAM_CONFIGURATION_NORMAL_MODE};
+      streams.size(), streams.data(), CAMERA3_STREAM_CONFIGURATION_NORMAL_MODE};
   PrepareDefaultDUT();
   EXPECT_FALSE(dut_->StreamConfigurationSupported(&config));
 }
